Feat/1924: Reward details API queries refactor. Also now shows rewards of any asset type
