+++ apps/token/src/routes/rewards/home/__generated___/Rewards.ts
@@ -3,46 +3,60 @@ import { Schema as Types } from '@vegaprotocol/types';
 import { gql } from '@apollo/client';
 import * as Apollo from '@apollo/client';
 const defaultOptions = {} as const;
+export type RewardFieldsFragment = { __typename?: 'Reward', rewardType: Types.AccountType, amount: string, amountFormatted: string, percentageOfTotal: string, receivedAt: string, asset: { __typename?: 'Asset', id: string, symbol: string }, party: { __typename?: 'Party', id: string }, epoch: { __typename?: 'Epoch', id: string } };
+
+export type DelegationFieldsFragment = { __typename?: 'Delegation', amount: string, amountFormatted: string, epoch: number };
+
 export type RewardsQueryVariables = Types.Exact<{
   partyId: Types.Scalars['ID'];
 }>;
 
 
-export type RewardsQuery = { __typename?: 'Query', party?: { __typename?: 'Party', id: string, rewardDetails?: Array<{ __typename?: 'RewardPerAssetDetail', totalAmount: string, totalAmountFormatted: string, asset: { __typename?: 'Asset', id: string, symbol: string }, rewards?: Array<{ __typename?: 'Reward', rewardType: Types.AccountType, amount: string, amountFormatted: string, percentageOfTotal: string, receivedAt: string, asset: { __typename?: 'Asset', id: string }, party: { __typename?: 'Party', id: string }, epoch: { __typename?: 'Epoch', id: string } } | null> | null } | null> | null, delegations?: Array<{ __typename?: 'Delegation', amount: string, amountFormatted: string, epoch: number }> | null } | null, epoch: { __typename?: 'Epoch', id: string, timestamps: { __typename?: 'EpochTimestamps', start?: string | null, end?: string | null, expiry?: string | null } } };
+export type RewardsQuery = { __typename?: 'Query', party?: { __typename?: 'Party', id: string, rewardsConnection?: { __typename?: 'RewardsConnection', edges?: Array<{ __typename?: 'RewardEdge', node: { __typename?: 'Reward', rewardType: Types.AccountType, amount: string, amountFormatted: string, percentageOfTotal: string, receivedAt: string, asset: { __typename?: 'Asset', id: string, symbol: string }, party: { __typename?: 'Party', id: string }, epoch: { __typename?: 'Epoch', id: string } } } | null> | null } | null, delegationsConnection?: { __typename?: 'DelegationsConnection', edges?: Array<{ __typename?: 'DelegationEdge', node: { __typename?: 'Delegation', amount: string, amountFormatted: string, epoch: number } } | null> | null } | null } | null, epoch: { __typename?: 'Epoch', id: string, timestamps: { __typename?: 'EpochTimestamps', start?: string | null, end?: string | null, expiry?: string | null } } };
 
-
+export const RewardFieldsFragmentDoc = gql`
+    fragment RewardFields on Reward {
+  rewardType
+  asset {
+    id
+    symbol
+  }
+  party {
+    id
+  }
+  epoch {
+    id
+  }
+  amount
+  amountFormatted @client
+  percentageOfTotal
+  receivedAt
+}
+    `;
+export const DelegationFieldsFragmentDoc = gql`
+    fragment DelegationFields on Delegation {
+  amount
+  amountFormatted @client
+  epoch
+}
+    `;
 export const RewardsDocument = gql`
     query Rewards($partyId: ID!) {
   party(id: $partyId) {
     id
-    rewardDetails {
-      asset {
-        id
-        symbol
+    rewardsConnection {
+      edges {
+        node {
+          ...RewardFields
+        }
       }
-      rewards {
-        rewardType
-        asset {
-          id
-        }
-        party {
-          id
-        }
-        epoch {
-          id
-        }
-        amount
-        amountFormatted @client
-        percentageOfTotal
-        receivedAt
-      }
-      totalAmount
-      totalAmountFormatted @client
     }
-    delegations {
-      amount
-      amountFormatted @client
-      epoch
+    delegationsConnection {
+      edges {
+        node {
+          ...DelegationFields
+        }
+      }
     }
   }
   epoch {
@@ -54,7 +68,8 @@ export const RewardsDocument = gql`
     }
   }
 }
-    `;
+    ${RewardFieldsFragmentDoc}
+${DelegationFieldsFragmentDoc}`;
 
 /**
  * __useRewardsQuery__
